+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.0.2
+
+* The method `BodyBuilder#_hasMore()` is now public: `BodyBuilder#hasMore()`
+* Add parameter `useButton`, `loadMoreBuilder` and `loadMoreLabel` to the widget `LoadMore` to allow
+  more customization of the load more button.
+
 ## 1.0.1
 
 * Updated the `README.md`to fix the broken links